+++ src/main/java/com/github/japgolly/colourmath/ColourHSL01.java
@@ -15,9 +15,9 @@ import lombok.experimental.Wither;
 @ToString
 @EqualsAndHashCode(callSuper = false)
 public class ColourHSL01 extends ColourAdapter {
-	@Wither public final float h, s, l;
+	@Wither public final double h, s, l;
 
-	ColourHSL01(float h, float s, float l) {
+	ColourHSL01(double h, double s, double l) {
 		Conditions.assert01(h, s, l);
 		this.h = h;
 		this.s = s;
@@ -31,39 +31,39 @@ public class ColourHSL01 extends ColourAdapter {
 
 	@Override
 	public ColourHSL255 HSL255() {
-		return new ColourHSL255((int) (h * 255f), (int) (s * 255f), (int) (l * 255f));
+		return new ColourHSL255((int) (h * 255.), (int) (s * 255.), (int) (l * 255.));
 	}
 
 	@Override
 	public ColourRGB01 RGB01() {
-		final float r, g, b;
-		if (s == 0f) {
+		final double r, g, b;
+		if (s == 0.) {
 			r = g = b = l; // achromatic
 		} else {
-			float q = l < 0.5f ? l * (1 + s) : l + s - l * s;
-			float p = 2f * l - q;
-			r = hue2rgb(p, q, h + 1f / 3f);
+			double q = l < 0.5 ? l * (1 + s) : l + s - l * s;
+			double p = 2. * l - q;
+			r = hue2rgb(p, q, h + 1. / 3.);
 			g = hue2rgb(p, q, h);
-			b = hue2rgb(p, q, h - 1f / 3f);
+			b = hue2rgb(p, q, h - 1. / 3.);
 		}
 		return new ColourRGB01(r, g, b);
 	}
 
-	private static float hue2rgb(float p, float q, float t) {
-		if (t < 0f) {
-			t += 1f;
+	private static double hue2rgb(double p, double q, double t) {
+		if (t < 0.) {
+			t += 1.;
 		}
-		if (t > 1f) {
-			t -= 1f;
+		if (t > 1.) {
+			t -= 1.;
 		}
-		if (t < 1f / 6f) {
-			return p + (q - p) * 6f * t;
+		if (t < 1. / 6.) {
+			return p + (q - p) * 6. * t;
 		}
-		if (t < 1f / 2f) {
+		if (t < 1. / 2.) {
 			return q;
 		}
-		if (t < 2f / 3f) {
-			return p + (q - p) * (2f / 3f - t) * 6f;
+		if (t < 2. / 3.) {
+			return p + (q - p) * (2. / 3. - t) * 6.;
 		}
 		return p;
 	}
